Delete schemautil
