+++ src/util.ts
@@ -73,6 +73,37 @@ export function all<T>(arr: Array<T>, f: (d: T, k?, i?) => boolean) {
   return true;
 }
 
+export function mergeDeep(dest, ...src: any[]) {
+  for (var i = 0; i < src.length; i++) {
+    dest = deepMerge_(dest, src[i]);
+  }
+  return dest;
+};
+
+// recursively merges src into dest
+function deepMerge_(dest, src) {
+  if (typeof src !== 'object' || src === null) {
+    return dest;
+  }
+
+  for (var p in src) {
+    if (!src.hasOwnProperty(p)) {
+      continue;
+    }
+    if (src[p] === undefined) {
+      continue;
+    }
+    if (typeof src[p] !== 'object' || src[p] === null) {
+      dest[p] = src[p];
+    } else if (typeof dest[p] !== 'object' || dest[p] === null) {
+      dest[p] = mergeDeep(src[p].constructor === Array ? [] : {}, src[p]);
+    } else {
+      mergeDeep(dest[p], src[p]);
+    }
+  }
+  return dest;
+}
+
 // FIXME remove this
 import * as dlBin from 'datalib/src/bins/bins';
 export function getbins(stats, maxbins) {
