+++ tsconfig.json
@@ -69,7 +69,6 @@
     "src/schema/legend.schema.ts",
     "src/schema/scale.schema.ts",
     "src/schema/schema.ts",
-    "src/schema/schemautil.ts",
     "src/schema/sort.schema.ts",
     "src/schema/transform.schema.ts",
     "src/schema/vega.schema.ts",
